fix: rename `SSelectItem` to `SSelectOption`
