+++ packages/ui/src/components/SSelect/index.ts
@@ -6,10 +6,10 @@ export { default as SDropdown } from './SDropdown.vue'
 export { default as SSelectBase } from './SSelectBase.vue'
 export { default as SSelectButton } from './SSelectButton.vue'
 export { default as SSelectInput } from './SSelectInput.vue'
-export { default as SSelectItem } from './SSelectItem.vue'
+export { default as SSelectOption } from './SSelectOption.vue'
 export { default as SSelectChevron } from './SSelectChevron'
 
 // tools
-export { SSelectSize, SSelectButtonType, SSelectItemType } from './types'
+export { SSelectSize, SSelectButtonType, SSelectOptionType } from './types'
 export { useSSelectApi } from './api'
 export type { SSelectApi } from './api'
